+++ src/components/Video/VideoFunctions.tsx
@@ -1,0 +1,21 @@
+export function getId(url) {
+  console.log("HELLO")
+  const regExp = /^.*(youtu.be\/|v\/|u\/\w\/|embed\/|watch\?v=|&v=)([^#&?]*).*/;
+  const match = url.match(regExp);
+  return (match && match[2].length === 11) ? match[2] : null;
+}
+
+export function arrayBufferToBlob(buffer, type) {
+  return new Blob([buffer], { type: type });
+}
+
+export function blobToArrayBuffer(blob) {
+  return new Promise((resolve, reject) => {
+    const reader = new FileReader();
+    reader.addEventListener('loadend', (e) => {
+      resolve(reader.result);
+    });
+    reader.addEventListener('error', reject);
+    reader.readAsArrayBuffer(blob);
+  });
+}
