+++ src/components/Video/VideoStyles.tsx
@@ -1,0 +1,8 @@
+import { createStyles, Theme, fade } from '@material-ui/core/styles';
+export const styles = (theme: Theme) =>
+  createStyles({
+    root: {
+      flexGrow: 1,
+      width: '100%',
+    },
+  });
